Update README.md
Updated documentation to reflect the proper way to start the container if the host is running Docker Desktop as opposed to Docker Engine. Docker Desktop cannot use the --net=host option and so as a result will fail to listen on any ports. It will post but will never connect, which is kinda frustrating. Using the -p option allows you to connect and use the node