Blender Download URL changed
The Blender download URL was specified as a separate input to the github action file. This is changed so that now the URL is constructed from other variables, namely the short and long version string, allowing you to customise the version of Blender used.

I also removed a build arg from the make file as it was overriding the version number with nothing causing a build error.