removed uneeded options
Only need the filepath to the addon package now, other options have been removed.

Support for using a copy of the addon that is on the host, but not in the container is done by simply omitting the -i flag, and making sure the local copy is bound to the container using docker's bind mounts.