Removed unecessary args
The sript called for both an image name and a tag, its simpler to just accept the image name with the tag included in it.

Removed checking for this tag as well.

Added a variable to hold the path ot the location the addon should be installed when running a dev session. All dev sessions use a copy of the addon hosted on the host, not in the container, hence the use of a bind mount.